langgraph-multiagent-skeleton/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-multiagent-skeleton"
version = "0.1.0"
description = "Starter skeleton for multi-agent LangGraph applications"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "pydantic>=2.0",
    "streamlit>=1.35",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
